+++ src/layer/marker/Marker.js
@@ -65,9 +65,13 @@ L.Marker = L.Class.extend({
 	_initInteraction: function() {
 		if (this.options.clickable) {
 			this._icon.className += ' leaflet-clickable';
-			L.DomEvent.addListener(this._icon, 'mousedown', this._fireMouseEvent, this);
+			
 			L.DomEvent.addListener(this._icon, 'click', this._onMouseClick, this);
-			L.DomEvent.addListener(this._icon, 'dblclick', this._fireMouseEvent, this);
+
+			var events = ['dblclick', 'mousedown', 'mouseover', 'mouseout'];
+			for (var i = 0; i < events.length; i++) {
+				L.DomEvent.addListener(this._icon, events[i], this._fireMouseEvent, this);
+			}
 		}
 		
 		if (L.Handler.MarkerDrag) {